Move external libs to "external" folder
